+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "hw.dt83.udpchat"
+        applicationId "com.application.callAuth"
         minSdkVersion 16
         targetSdkVersion 26
     }
@@ -20,5 +20,5 @@ android {
 
 dependencies {
     implementation project(path: ':recognito-0.1.0-SNAPSHOT')
-    compile 'com.android.support:support-v4:18.0.0'
+    compile 'com.android.support:support-v4:26.1.0'
 }
