+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="hw.dt83.udpchat"
+    package="com.application.callAuth"
     android:versionCode="1"
     android:versionName="1.0" >
 
@@ -16,7 +16,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="hw.dt83.udpchat.MainActivity"
+            android:name="com.application.callAuth.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait" >
             <intent-filter>
@@ -26,12 +26,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="hw.dt83.udpchat.MakeCallActivity"
+            android:name="com.application.callAuth.MakeCallActivity"
             android:label="@string/title_activity_make_call"
             android:screenOrientation="portrait" >
         </activity>
         <activity
-            android:name="hw.dt83.udpchat.ReceiveCallActivity"
+            android:name="com.application.callAuth.ReceiveCallActivity"
             android:label="@string/title_activity_receive_call"
             android:screenOrientation="portrait" >
         </activity>
